making progress on the Vocabularies form and table
